requirements.txt
PyMuPDF==1.23.21
sentence-transformers==2.5.1
faiss-cpu==1.8.0
Pillow==10.2.0
transformers==4.38.2
torch==2.2.1
requests==2.31.0
tqdm==4.66.2
pyyaml==6.0.1
pydantic>=2.9.0
numpy==1.26.4
langchain-text-splitters==0.0.1
google-generativeai==0.8.3
python-dotenv==1.0.1
